Make KDC "status" statements more homogeneous
Generally we want KDC status strings to be concise, informative and follow
some common rule. Many of the current status messages are in "DO_THIS" format,
i.e. uppercased string with underscore as a separator. For example,
"DECRYPT_SERVER_KEY".  This commit is to modify some KDC status messages to
follow this format.

Even though KDC status messages are not standardized, it is possible that some
administrators use them in the Kerberos log file processing. Hence, the vast
majority of them are left unchanged pending further investigation (mostly,
feedback from the administrators).